CountryStateCity

+++ app/Http/Controllers/DeleteCommentController.php
@@ -4,14 +4,17 @@ namespace App\Http\Controllers;
 
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class DeleteCommentController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Handle the incoming request.
      */
     public function __invoke($id,Comment $comment)
     {
+        $this->authorize('delete', $comment);
         $comment->delete();
         return back();
     }
